+++ chessboard/solver.py
@@ -159,7 +159,7 @@ class SolverContext(object):
                     # Translate linear index to 2D dimension.
                     x, y = board.index_to_coordinates(linear_position)
                     # Try to place the piece on the board.
-                    board.add(piece_kind, x, y)
+                    board.add(piece_kind, linear_position)
                 # If one of the piece can't be added, throw the whole set, skip
                 # the rotten branch and proceed to the next.
                 except (OccupiedPosition, VulnerablePosition, AttackablePiece):