+++ chessboard/tests/test_pieces.py
@@ -36,47 +36,49 @@ class TestKing(unittest.TestCase):
 
     def test_territory(self):
         """ Test computation of territory at each positions of a 3x3 board. """
-        self.assertEquals(King(Board(3, 3), 1, 1).territory, [
+        board = Board(3, 3)
+        c2i = board.coordinates_to_index
+        self.assertEquals(King(board, c2i(1, 1)).territory, [
              True,  True,  True,
              True,  True,  True,
              True,  True,  True,
         ])
-        self.assertEquals(King(Board(3, 3), 0, 0).territory, [
+        self.assertEquals(King(board, c2i(0, 0)).territory, [
              True,  True, False,
              True,  True, False,
             False, False, False,
         ])
-        self.assertEquals(King(Board(3, 3), 1, 0).territory, [
+        self.assertEquals(King(board, c2i(1, 0)).territory, [
              True,  True,  True,
              True,  True,  True,
             False, False, False,
         ])
-        self.assertEquals(King(Board(3, 3), 2, 0).territory, [
+        self.assertEquals(King(board, c2i(2, 0)).territory, [
             False,  True,  True,
             False,  True,  True,
             False, False, False,
         ])
-        self.assertEquals(King(Board(3, 3), 2, 1).territory, [
+        self.assertEquals(King(board, c2i(2, 1)).territory, [
             False,  True,  True,
             False,  True,  True,
             False,  True,  True,
         ])
-        self.assertEquals(King(Board(3, 3), 2, 2).territory, [
+        self.assertEquals(King(board, c2i(2, 2)).territory, [
             False, False, False,
             False,  True,  True,
             False,  True,  True,
         ])
-        self.assertEquals(King(Board(3, 3), 1, 2).territory, [
+        self.assertEquals(King(board, c2i(1, 2)).territory, [
             False, False, False,
              True,  True,  True,
              True,  True,  True,
         ])
-        self.assertEquals(King(Board(3, 3), 0, 2).territory, [
+        self.assertEquals(King(board, c2i(0, 2)).territory, [
             False, False, False,
              True,  True, False,
              True,  True, False,
         ])
-        self.assertEquals(King(Board(3, 3), 0, 1).territory, [
+        self.assertEquals(King(board, c2i(0, 1)).territory, [
              True,  True, False,
              True,  True, False,
              True,  True, False,
@@ -87,47 +89,49 @@ class TestQueen(unittest.TestCase):
 
     def test_territory(self):
         """ Test computation of territory at each positions of a 3x3 board. """
-        self.assertEquals(Queen(Board(3, 3), 1, 1).territory, [
+        board = Board(3, 3)
+        c2i = board.coordinates_to_index
+        self.assertEquals(Queen(board, c2i(1, 1)).territory, [
              True,  True,  True,
              True,  True,  True,
              True,  True,  True,
         ])
-        self.assertEquals(Queen(Board(3, 3), 0, 0).territory, [
+        self.assertEquals(Queen(board, c2i(0, 0)).territory, [
              True,  True,  True,
              True,  True, False,
              True, False,  True,
         ])
-        self.assertEquals(Queen(Board(3, 3), 1, 0).territory, [
+        self.assertEquals(Queen(board, c2i(1, 0)).territory, [
              True,  True,  True,
              True,  True,  True,
             False,  True, False,
         ])
-        self.assertEquals(Queen(Board(3, 3), 2, 0).territory, [
+        self.assertEquals(Queen(board, c2i(2, 0)).territory, [
              True,  True,  True,
             False,  True,  True,
              True, False,  True,
         ])
-        self.assertEquals(Queen(Board(3, 3), 2, 1).territory, [
+        self.assertEquals(Queen(board, c2i(2, 1)).territory, [
             False,  True,  True,
              True,  True,  True,
             False,  True,  True,
         ])
-        self.assertEquals(Queen(Board(3, 3), 2, 2).territory, [
+        self.assertEquals(Queen(board, c2i(2, 2)).territory, [
              True, False,  True,
             False,  True,  True,
              True,  True,  True,
         ])
-        self.assertEquals(Queen(Board(3, 3), 1, 2).territory, [
+        self.assertEquals(Queen(board, c2i(1, 2)).territory, [
             False,  True, False,
              True,  True,  True,
              True,  True,  True,
         ])
-        self.assertEquals(Queen(Board(3, 3), 0, 2).territory, [
+        self.assertEquals(Queen(board, c2i(0, 2)).territory, [
              True, False,  True,
              True,  True, False,
              True,  True,  True,
         ])
-        self.assertEquals(Queen(Board(3, 3), 0, 1).territory, [
+        self.assertEquals(Queen(board, c2i(0, 1)).territory, [
              True,  True, False,
              True,  True,  True,
              True,  True, False,
@@ -138,47 +142,49 @@ class TestRook(unittest.TestCase):
 
     def test_territory(self):
         """ Test computation of territory at each positions of a 3x3 board. """
-        self.assertEquals(Rook(Board(3, 3), 1, 1).territory, [
+        board = Board(3, 3)
+        c2i = board.coordinates_to_index
+        self.assertEquals(Rook(board, c2i(1, 1)).territory, [
             False,  True,  False,
              True,  True,  True,
             False,  True,  False,
         ])
-        self.assertEquals(Rook(Board(3, 3), 0, 0).territory, [
+        self.assertEquals(Rook(board, c2i(0, 0)).territory, [
              True,  True,  True,
              True, False, False,
              True, False, False,
         ])
-        self.assertEquals(Rook(Board(3, 3), 1, 0).territory, [
+        self.assertEquals(Rook(board, c2i(1, 0)).territory, [
              True,  True,  True,
             False,  True, False,
             False,  True, False,
         ])
-        self.assertEquals(Rook(Board(3, 3), 2, 0).territory, [
+        self.assertEquals(Rook(board, c2i(2, 0)).territory, [
              True,  True,  True,
             False, False,  True,
             False, False,  True,
         ])
-        self.assertEquals(Rook(Board(3, 3), 2, 1).territory, [
+        self.assertEquals(Rook(board, c2i(2, 1)).territory, [
             False, False,  True,
              True,  True,  True,
             False, False,  True,
         ])
-        self.assertEquals(Rook(Board(3, 3), 2, 2).territory, [
+        self.assertEquals(Rook(board, c2i(2, 2)).territory, [
             False, False,  True,
             False, False,  True,
              True,  True,  True,
         ])
-        self.assertEquals(Rook(Board(3, 3), 1, 2).territory, [
+        self.assertEquals(Rook(board, c2i(1, 2)).territory, [
             False,  True, False,
             False,  True, False,
              True,  True,  True,
         ])
-        self.assertEquals(Rook(Board(3, 3), 0, 2).territory, [
+        self.assertEquals(Rook(board, c2i(0, 2)).territory, [
              True, False, False,
              True, False, False,
              True,  True,  True,
         ])
-        self.assertEquals(Rook(Board(3, 3), 0, 1).territory, [
+        self.assertEquals(Rook(board, c2i(0, 1)).territory, [
              True, False, False,
              True,  True,  True,
              True, False, False,
@@ -189,47 +195,49 @@ class TestBishop(unittest.TestCase):
 
     def test_territory(self):
         """ Test computation of territory at each positions of a 3x3 board. """
-        self.assertEquals(Bishop(Board(3, 3), 1, 1).territory, [
+        board = Board(3, 3)
+        c2i = board.coordinates_to_index
+        self.assertEquals(Bishop(board, c2i(1, 1)).territory, [
              True, False,  True,
             False,  True, False,
              True, False,  True,
         ])
-        self.assertEquals(Bishop(Board(3, 3), 0, 0).territory, [
+        self.assertEquals(Bishop(board, c2i(0, 0)).territory, [
              True, False, False,
             False,  True, False,
             False, False,  True,
         ])
-        self.assertEquals(Bishop(Board(3, 3), 1, 0).territory, [
+        self.assertEquals(Bishop(board, c2i(1, 0)).territory, [
             False,  True, False,
              True, False,  True,
             False, False, False,
         ])
-        self.assertEquals(Bishop(Board(3, 3), 2, 0).territory, [
+        self.assertEquals(Bishop(board, c2i(2, 0)).territory, [
             False, False,  True,
             False,  True, False,
              True, False, False,
         ])
-        self.assertEquals(Bishop(Board(3, 3), 2, 1).territory, [
+        self.assertEquals(Bishop(board, c2i(2, 1)).territory, [
             False,  True, False,
             False, False,  True,
             False,  True, False,
         ])
-        self.assertEquals(Bishop(Board(3, 3), 2, 2).territory, [
+        self.assertEquals(Bishop(board, c2i(2, 2)).territory, [
              True, False, False,
             False,  True, False,
             False, False,  True,
          ])
-        self.assertEquals(Bishop(Board(3, 3), 1, 2).territory, [
+        self.assertEquals(Bishop(board, c2i(1, 2)).territory, [
             False, False, False,
              True, False,  True,
             False,  True, False,
         ])
-        self.assertEquals(Bishop(Board(3, 3), 0, 2).territory, [
+        self.assertEquals(Bishop(board, c2i(0, 2)).territory, [
             False, False,  True,
             False,  True, False,
              True, False, False,
         ])
-        self.assertEquals(Bishop(Board(3, 3), 0, 1).territory, [
+        self.assertEquals(Bishop(board, c2i(0, 1)).territory, [
             False,  True, False,
              True, False, False,
             False,  True, False,
@@ -240,47 +248,49 @@ class TestKnight(unittest.TestCase):
 
     def test_territory(self):
         """ Test computation of territory at each positions of a 3x3 board. """
-        self.assertEquals(Knight(Board(3, 3), 1, 1).territory, [
+        board = Board(3, 3)
+        c2i = board.coordinates_to_index
+        self.assertEquals(Knight(board, c2i(1, 1)).territory, [
             False, False, False,
             False,  True, False,
             False, False, False,
         ])
-        self.assertEquals(Knight(Board(3, 3), 0, 0).territory, [
+        self.assertEquals(Knight(board, c2i(0, 0)).territory, [
              True, False, False,
             False, False,  True,
             False,  True, False,
         ])
-        self.assertEquals(Knight(Board(3, 3), 1, 0).territory, [
+        self.assertEquals(Knight(board, c2i(1, 0)).territory, [
             False,  True, False,
             False, False, False,
              True, False,  True,
         ])
-        self.assertEquals(Knight(Board(3, 3), 2, 0).territory, [
+        self.assertEquals(Knight(board, c2i(2, 0)).territory, [
             False, False,  True,
              True, False, False,
             False,  True, False,
         ])
-        self.assertEquals(Knight(Board(3, 3), 2, 1).territory, [
+        self.assertEquals(Knight(board, c2i(2, 1)).territory, [
              True, False, False,
             False, False,  True,
              True, False, False,
         ])
-        self.assertEquals(Knight(Board(3, 3), 2, 2).territory, [
+        self.assertEquals(Knight(board, c2i(2, 2)).territory, [
             False,  True, False,
              True, False, False,
             False, False,  True,
          ])
-        self.assertEquals(Knight(Board(3, 3), 1, 2).territory, [
+        self.assertEquals(Knight(board, c2i(1, 2)).territory, [
              True, False,  True,
             False, False, False,
             False,  True, False,
         ])
-        self.assertEquals(Knight(Board(3, 3), 0, 2).territory, [
+        self.assertEquals(Knight(board, c2i(0, 2)).territory, [
             False,  True, False,
             False, False,  True,
              True, False, False,
         ])
-        self.assertEquals(Knight(Board(3, 3), 0, 1).territory, [
+        self.assertEquals(Knight(board, c2i(0, 1)).territory, [
             False, False,  True,
              True, False, False,
             False, False,  True,