+++ chessboard/tests/test_solver.py
@@ -27,6 +27,7 @@ import unittest
 from chessboard import (
     Permutations,
     SolverContext,
+    KING, QUEEN,
 )
 
 
@@ -368,7 +369,7 @@ class TestSolverContext(unittest.TestCase):
         solver = SolverContext(3, 3, king=2, queen=7)
         self.assertEquals(solver.length, 3)
         self.assertEquals(solver.height, 3)
-        self.assertDictContainsSubset({'king': 2, 'queen': 7}, solver.pieces)
+        self.assertDictContainsSubset({KING: 2, QUEEN: 7}, solver.pieces)
 
     def check_results(self, results, expected):
         """ Check found results.