+++ CHANGES.rst
@@ -5,7 +5,8 @@ ChangeLog
 0.9.2 (unreleased)
 ------------------
 
-* No changes yet.
+* Do not spend time converting back and forth linear position to 2D position.
+  Provides a ~5% speed-up.
 
 
 0.9.1 (2015-08-25)